fix various bugs

+++ src/redshift.c
@@ -63,7 +63,7 @@ char config_time_format[REDSHIFT_STRINGCONFIG_MAXLEN+1] = "%I:0%M";
 char config_info_below[REDSHIFT_STRINGCONFIG_MAXLEN+1] = "%A, %m/%d";
 uint8_t config_update_second = 0;
 uint8_t config_show_daynight = true;
-uint8_t config_step_goal = 10000;
+uint16_t config_step_goal = 10000;
 // -- end autogen
